+++ packages/flame_tiled/lib/src/renderable_layers/tile_layers/orthogonal_tile_layer.dart
@@ -2,12 +2,12 @@ import 'package:flame/extensions.dart';
 import 'package:flame_tiled/flame_tiled.dart';
 import 'package:flame_tiled/src/mutable_rect.dart';
 import 'package:flame_tiled/src/mutable_transform.dart';
-import 'package:flame_tiled/src/renderable_layers/tile_layer/tile_layer.dart';
+import 'package:flame_tiled/src/renderable_layers/tile_layers/tile_layer.dart';
 import 'package:meta/meta.dart';
 
 @internal
-class IsometricTileLayer extends FlameTileLayer {
-  IsometricTileLayer({
+class OrthogonalTileLayer extends FlameTileLayer {
+  OrthogonalTileLayer({
     required super.layer,
     required super.parent,
     required super.map,
@@ -20,10 +20,7 @@ class IsometricTileLayer extends FlameTileLayer {
   @override
   void cacheTiles() {
     final tileData = layer.tileData!;
-    final halfDestinationTile = destTileSize / 2;
     final size = destTileSize;
-    final isometricXShift = map.height * halfDestinationTile.x;
-    final isometricYShift = halfDestinationTile.y;
     final halfMapTile = Vector2(map.tileWidth / 2, map.tileHeight / 2);
     final batch = tiledAtlas.batch;
     if (batch == null) {
@@ -42,6 +39,7 @@ class IsometricTileLayer extends FlameTileLayer {
         final tile = map.tileByGid(tileGid.tile)!;
         final tileset = map.tilesetByTileGId(tileGid.tile);
         final img = tile.image ?? tileset.image;
+
         if (img == null) {
           continue;
         }
@@ -57,6 +55,7 @@ class IsometricTileLayer extends FlameTileLayer {
               .toRect()
               .translate(spriteOffset.dx, spriteOffset.dy),
         );
+
         final flips = SimpleFlips.fromFlips(tileGid.flips);
         final scale = size.x / src.width;
         final anchorX = src.width - halfMapTile.x;
@@ -64,9 +63,8 @@ class IsometricTileLayer extends FlameTileLayer {
 
         late double offsetX;
         late double offsetY;
-
-        offsetX = halfDestinationTile.x * (tx - ty) + isometricXShift;
-        offsetY = halfDestinationTile.y * (tx + ty) + isometricYShift;
+        offsetX = (tx + .5) * size.x;
+        offsetY = (ty + .5) * size.y;
 
         final scos = flips.cos * scale;
         final ssin = flips.sin * scale;
