+++ packages/flame_tiled/lib/src/renderable_layers/tile_layers/tile_layer.dart
@@ -5,10 +5,10 @@ import 'package:flame_tiled/src/mutable_rect.dart';
 import 'package:flame_tiled/src/mutable_transform.dart';
 import 'package:flame_tiled/src/renderable_layers/group_layer.dart';
 import 'package:flame_tiled/src/renderable_layers/renderable_layer.dart';
-import 'package:flame_tiled/src/renderable_layers/tile_layer/hexagonal_tile_layer.dart';
-import 'package:flame_tiled/src/renderable_layers/tile_layer/isometric_tile_layer.dart';
-import 'package:flame_tiled/src/renderable_layers/tile_layer/orthogonal_tile_layer.dart';
-import 'package:flame_tiled/src/renderable_layers/tile_layer/staggered_tile_layer.dart';
+import 'package:flame_tiled/src/renderable_layers/tile_layers/hexagonal_tile_layer.dart';
+import 'package:flame_tiled/src/renderable_layers/tile_layers/isometric_tile_layer.dart';
+import 'package:flame_tiled/src/renderable_layers/tile_layers/orthogonal_tile_layer.dart';
+import 'package:flame_tiled/src/renderable_layers/tile_layers/staggered_tile_layer.dart';
 import 'package:flame_tiled/src/tile_animation.dart';
 import 'package:flame_tiled/src/tile_atlas.dart';
 import 'package:flutter/painting.dart';
